requirements.txt
pkg-resources==0.0.0
pyelftools==0.24
